+++ RELEASE_NOTES.md
@@ -8,6 +8,7 @@ Changes:
     tissue piece children will automatically be deducted from the slides' (current) slide count on
     save. This update will use the current slide count as initial slides, and set the current slide
     count to (initial slides - discards - slides consumed by all children)
+  * Show 1 decimal place on labels for volume/concentration values less than 10
 
 # 0.2.192